We launch single item details in a new screen
We launch single item details in a new screen inside the onCreate method.
Which means that this method nows: get data, connect the data to the
listview and finally checks whether someone clicks on one item. Once it's
been clicked, create a new intent and send the data to this new activity

NEXT: Do something with the data we are sending to the activity

Contact: joaquin@isla.io